\name{longclust-package}
\alias{longclust-package}
\alias{longclust}
\docType{package}
\title{Model-Based Clustering and Classification for Longitudinal Data}
\description{
This is a package for clustering or classification of longitudinal data based on a mixture of multivariate t or Gaussian distributions with a Cholesky-decomposed covariance structure.
}
\details{
\tabular{ll}{
Package: \tab longclust\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-05-28\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
This package contains the function \code{\link{longclustEM}}.
}
\author{
P. D. McNicholas, K.R. Jampani and S. Subedi

Maintainer: K. Raju Jampani <rjampani@uoguelph.ca>
}
\seealso{
Details, examples, and references are given under \code{\link{longclustEM}}.
}